+++ R/makeGRANRepos.R
@@ -1,7 +1,7 @@
 #' @rdname makerepo
 #' @aliases makeRepo,PkgManifest
 setMethod("makeRepo", "PkgManifest",
-          function(x, cores = (parallel:::detectCores() - 1), build_pkgs = NULL,
+          function(x, cores = 1, build_pkgs = NULL,
                    scm_auth = list("bioconductor.org" =
                        c("readonly", "readonly")),
                    ...
@@ -22,7 +22,7 @@ setMethod("makeRepo", "PkgManifest",
 #' @aliases makeRepo,SessionManifest
 
 setMethod("makeRepo", "SessionManifest",
-          function(x, cores = (parallel:::detectCores() - 1), build_pkgs = NULL,
+          function(x, cores = 1, build_pkgs = NULL,
                    scm_auth = list("bioconductor.org" =
                        c("readonly", "readonly")),
                    ...
@@ -39,7 +39,7 @@ setMethod("makeRepo", "SessionManifest",
 #' @rdname makerepo
 #' @aliases makeRepo,GRANRepository
 setMethod("makeRepo", "GRANRepository",
-          function(x, cores = (parallel:::detectCores() - 1), build_pkgs = NULL,
+          function(x, cores = 1, build_pkgs = NULL,
                    scm_auth = list("bioconductor.org" =
                                        c("readonly", "readonly")),
                    ...) {
@@ -51,12 +51,12 @@ setMethod("makeRepo", "GRANRepository",
     }
     repo = x
     if(file.exists(destination(repo)))
-        repo2 = tryCatch(loadRepo(paste(destination(repo), "repo.R",
-                                        sep="/")), error = function(x) NULL)
+        repo2 = suppressWarnings(tryCatch(
+                    loadRepo(file.path(destination(repo), "repo.R")), error = function(x) NULL))
     else
         repo2 = suppressWarnings(tryCatch(
-                              loadRepo(paste(repo_url(repo), "repo.R", sep="/")), 
-                              error = function(x) NULL))
+                    loadRepo(paste(repo_url(repo), "repo.R", sep="/")),
+                    error = function(x) NULL))
     if(!is.null(repo2) ) {
         res = repo_results(repo)
         res2 = repo_results(repo2)
@@ -105,6 +105,11 @@ setMethod("makeRepo", "GRANRepository",
     message(paste("Built", sum(getBuilding(repo)), "packages"))
     repo = suppressWarnings(migrateToFinalRepo(repo))
 
+    # Create Windows binaries
+    if (make_windows_bins(repo) && getOS() == "windows") {
+        makeWinBins(repo, cores = cores)
+    }
+
     finalizeRepo(repo)
     message(paste("Completed makeRepo at", Sys.time()))
     return(repo)
@@ -114,7 +119,7 @@ setMethod("makeRepo", "GRANRepository",
 #' @aliases makeRepo,character
 
 setMethod("makeRepo", "character",
-          function(x, cores = (parallel:::detectCores() - 1), build_pkgs = NULL,
+          function(x, cores = 1, build_pkgs = NULL,
                    scm_auth = list("bioconductor.org" =
                        c("readonly", "readonly")),
                    ...) {