+++ NEWS.md
@@ -1,3 +1,74 @@
+Changes in version 2.4.1 (2017-04-25)
+
+- Bugfixes
+  * Use relative paths for coverage report hyperlinks in the build report and package docs
+
+Changes in version 2.4.0 (2017-04-23)
+
+- New Features
+  * Add link to view package metadata JSON in package docs
+
+- Bugfixes
+  * Use relative paths instead of absolute for hyperlinks in package docs
+
+Changes in version 2.3.1 (2017-04-12)
+
+- Bugfixes
+  * Change label for package archives
+  * Not saving splash pages for previous versions of the package
+
+Changes in version 2.3.0 (2017-04-11)
+
+- API Changes
+  * createJSON now has an additional param for rev_deps
+
+- New Features
+  * Package metadata JSON: Add reverse dependency info and convert deps to JSON array
+
+Changes in version 2.2.0 (2017-04-05)
+
+- New Features
+  * Add hyperlinks for package source build and archive
+  * Save previous splash pages instead of always overwriting the existing one
+
+- API Changes
+  * Renamed createURL() to createHyperlink()
+
+- Improvements
+  * Use generic pkg_doc_dir instead of hardcoding package doc dir location
+
+Changes in version 2.1.0 (2018-04-03)
+
+- API Changes
+  * manifestHTML() is now called buildReport()
+
+- Improvements
+  * Always update packages in the temporary library before testing
+
+- Bugfixes
+  * Set correct statuses for package builds
+
+- New Features
+  * Added support for Windows binary builds
+  * Create an HTML report for the package manifest
+  * Metadata RDS files are created in the Meta folder of the repository
+  * Package README file links are available on package splash pages
+
+Changes in version 2.0.1 (2017-12-06)
+
+- Bugfixes
+  * switchr dependency is versioned (>= 0.9.28)
+  * Specify vignette engine
+
+Changes in version 2.0.0 (2017-12-05)
+
+- API Changes
+  * GRANRepository classes and methods have moved to the GRANCore package
+
+- Improvements
+  * Vignette is now in the Rmd format, and does not rely on BiocStyle anymore
+  * GRAN repository packages are now smaller in size, and only depend on GRANCore
+
 Changes in version 1.6.7 (2017-12-04)
 
 - Improvements
@@ -11,7 +82,6 @@ Changes in version 1.6.7 (2017-12-04)
 Changes in version 1.6.6 (2017-11-10)
 
 - New Features
-  * Parallelize creation of covr reports
   * Create a CRAN-style archive directory, location can be optionally specified in repo params
   * Performance improvements
 
@@ -19,6 +89,15 @@ Changes in version 1.6.6 (2017-11-10)
   * Don't create covr reports if check_test is off
   * Handle cases when all packages build successfully
   * Risk report is properly styled
+
+Changes in version 1.6.5 (2017-11-02)
+
+- New Features
+  * Added README.md page
+
+- Improvements
+  * Force switchr to build only if remote (Github) version is greater than repo version
+  * Parallelize creation of covr reports
 
 Changes in version 1.6.4 (2017-09-01)
 
@@ -46,6 +125,7 @@ Changes in version 1.6.0 (2017-08-02)
 
 - New Features
   * Risk reports, build badges for covr reports & build status
+  * A JSON file containing metadata for the package is created in the package docs directory
   * Beautified build report
 
 Changes in version 1.5.0 (2017-07-18)